+++ lib/blazer/adapters/presto_adapter.rb
@@ -40,7 +40,8 @@ module Blazer
         @client ||= begin
           uri = URI.parse(settings["url"])
           query = uri.query ? CGI.parse(uri.query) : {}
-          Presto::Client.new(
+          cls = uri.scheme == "trino" ? Trino::Client : Presto::Client
+          cls.new(
             server: "#{uri.host}:#{uri.port}",
             catalog: uri.path.to_s.delete_prefix("/"),
             schema: query["schema"] || "public",